requirements.txt
django
djangorestframework
djangorestframework-jwt

